params: change blob count types to int
{Latest,}MaxBlobsPerBlock returns a length, which is traditionally an int in Go. When
comparing with blob gas (uint64) a conversion is required, but I think it's actually
helpful. Kind of makes it clearer that you're working with different units.
When comparing against a slice length, no conversion is needed.

I'm also adding MaxBlobGasPerBlock to simplify the most common expression involving
MaxBlobsPerBlock.

+++ consensus/misc/eip4844/eip4844_test.go
@@ -29,11 +29,11 @@ func TestCalcExcessBlobGas(t *testing.T) {
 	var (
 		config        = params.MainnetChainConfig
 		targetBlobs   = config.TargetBlobsPerBlock(*config.CancunTime)
-		targetBlobGas = targetBlobs * params.BlobTxBlobGasPerBlob
+		targetBlobGas = uint64(targetBlobs) * params.BlobTxBlobGasPerBlob
 	)
 	var tests = []struct {
 		excess uint64
-		blobs  uint64
+		blobs  int
 		want   uint64
 	}{
 		// The excess blob gas should not increase from zero if the used blob
@@ -56,7 +56,7 @@ func TestCalcExcessBlobGas(t *testing.T) {
 		{params.BlobTxBlobGasPerBlob - 1, targetBlobs - 1, 0},
 	}
 	for i, tt := range tests {
-		blobGasUsed := tt.blobs * params.BlobTxBlobGasPerBlob
+		blobGasUsed := uint64(tt.blobs) * params.BlobTxBlobGasPerBlob
 		parent := &types.Header{
 			Time:          *config.CancunTime,
 			ExcessBlobGas: &tt.excess,
