+++ core/txpool/validation.go
@@ -134,7 +134,7 @@ func ValidateTransaction(tx *types.Transaction, head *types.Header, signer types
 		if len(hashes) == 0 {
 			return errors.New("blobless blob transaction")
 		}
-		if max := opts.Config.MaxBlobsPerBlock(head.Time); uint64(len(hashes)) > max {
+		if max := opts.Config.MaxBlobsPerBlock(head.Time); len(hashes) > max {
 			return fmt.Errorf("too many blobs in transaction: have %d, permitted %d", len(hashes), max)
 		}
 		// Ensure commitments, proofs and hashes are valid
